% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check.r
\name{arg_check}
\alias{arg_check}
\title{Arguments checks}
\usage{
arg_check(x, mode, y)
}
\arguments{
\item{x}{a char array of codons}

\item{mode}{a char array denoting the expected mode of the argument x}

\item{y}{a number denoting the codon table to be used}
}
\value{
a numeric matrix of appropriate codon table
}
\description{
Performs the usual checks of arguments consistency of codon sequences and codon table to be used
}
