% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augc_freq.r
\name{augc_freq}
\alias{augc_freq}
\title{Overall AT and GC content}
\usage{
augc_freq(x)
}
\arguments{
\item{x}{a list of KZsqns objects}
}
\value{
a numeric matrix of AU and GC content
}
\description{
Calculate the overall AT and GC content
}
