% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cai.r
\name{cai}
\alias{cai}
\title{Codon adaptation index}
\source{
The reference set of very highly expressed genes for \emph{E. coli}, \emph{B. subtilis} and \emph{S. cerevisiae} from taken from the C program condonW version 1.4.4 writen by John Peden available at http://codonw.sourceforge.net/. For more information see, the documentation of condonW.
}
\usage{
cai(x, ref_spp = "n", ref, y = 0, threthold = 0.005,
  lower_bound = 0.005)
}
\arguments{
\item{x}{a char array of codons, for which CAI will be calculated}

\item{ref_spp}{a char sequence of species name to use. "ecoli" = \emph{Escherichia coli}; "bsubt" = \emph{Bacillus subtilis}; "scere" = \emph{Saccharomyces cerevisiae}. When specified, it overides the value set by \emph{ref}}

\item{ref}{a named numeric array of relative adaptiveness of each codon from a species-specific reference set of very highly expressed genes. The array should have a length of 64 with 0 filled in at the stop codons. The name of each number is the corresponding char array of codon. For example, "AAA" or "Aaa" represent the same codon. Upper case letters can be mixed with lower case letters.}

\item{y}{an optional number denoting the codon table to be used. 0 = standard codon table (default), 2 = vertibrate mitchodrial, 3 = ?, 4 = ?. When an option other than those mentioned above is provided, the standard codon table will be used}

\item{threthold}{a number below which the adaptiveness will be forced to \emph{lower_bound}}

\item{lower_bound}{a number to which values of adaptiveness below the \emph{threshold} will be forced}
}
\value{
a number denoting the CAI for the provided sequence
}
\description{
Calculates the codon adaptation index (CAI) for each codon
}
\details{
When any reference codon has a relative usage value of zero and the codon sequence to be calculated has such a codon, the CAI for this sequence would be
}
\examples{

}
