% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cai_crt.r
\name{cai_crt}
\alias{cai_crt}
\title{Custom CAI reference table}
\usage{
cai_crt(x, y = 0, method = c("sharp", "bulmer"))
}
\arguments{
\item{x}{a list of KZsqns object, from which the reference table will be calculated}

\item{y}{y an optional number denoting the codon table to be used. 0 = standard codon table (default), 2 = vertibrate mitchodrial (See dna2aa for additional options). When an option other than those mentioned above is provided, the standard codon table will be used}

\item{method}{a charactor string denoting which method to use for codons absent from the reference set: "bulmer", set the adaptiveness of the codon to 0.01; "sharp", in calculating the adaptiveness of an absent codon, use a frequency of 0.5 instead}
}
\description{
A helper method to calculate the codon adaptiveness index reference table based on a list of highly expressed genes
}
\details{
If any codon is missing from the reference set, the adaptiveness value of that codon need to be adjusted in order to obtain valid CAI for sequences that have that codon in their sequence. The sharp's method is the default option [1], but the user may choose to use bulmer's method [2].
}
\references{
[1] Sharp, P. M., & Li, W.-H. (1987). The codon adaptation index-a measure of directional synonymous codon usage bias, and its potential applications. Nucleic acids research, 15(3), 1281-1295.
[2] Bulmer, M. (1988). Are codon usage patterns in unicellular organisms determined by selection-mutation balance? Journal of Evolutionary Biology, 1(1), 15-26.
}
