% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enc_exp.r
\name{enc_exp}
\alias{enc_exp}
\title{Expected number of ENC given GC3 content}
\usage{
enc_exp(gc3 = 0)
}
\arguments{
\item{gc3}{a number of GC3 content}
}
\value{
a number of the expected ENC given the GC3 content
}
\description{
Calculates the expected ENC based on Wright's formula given GC3 percentage
}
