% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc3.r
\name{gc3}
\alias{gc3}
\title{Calculate GC3 value of a codon sequence}
\usage{
gc3(x)
}
\arguments{
\item{x}{a list of KZsqns objects.}
}
\value{
a number array of GC3s.
}
\description{
Outputs an array of numbers, representing the proportion of codons with G/C bases in the third position
}
\examples{
data('CodonTable0')
x = vector('list', 5) # Creating an empty list of length 5
for(i in 1:5){
  x[[i]] = CodonTable0[sample(1:64, 10*i, TRUE),1]
  attr(x[[i]], 'class') = 'KZsqns'
}

cat(gc3(x))
}
