% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcx.r
\name{gcx}
\alias{gcx}
\title{GCx of a gene}
\usage{
gcx(x, pos = 1)
}
\arguments{
\item{x}{a char array of codons}

\item{pos}{the position of the codon for which GCx is calculated. 1: GC1; 2:GC2; 3: GC3; for any other values, the overall GC content will be calculated with a warning}
}
\value{
a number of the GCx of the gene
}
\description{
Calculates the GC1,2,3 and the overall GC content of a gene
}
\examples{

}
