% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n3_freq.r
\name{n3_freq}
\alias{n3_freq}
\title{Nucleotide at the third position of the codon}
\usage{
n3_freq(x)
}
\arguments{
\item{x}{a list of KZsqns objects.}
}
\value{
a matrix of nucleotide composition at the third position
}
\description{
Calculates the nucleotide frequency at the third position of the codon
}
