% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_fuglsang4m.r
\name{nc_fuglsang4m}
\alias{nc_fuglsang4m}
\title{Effective number of codons based on Fuglsang's formula 4-}
\usage{
nc_fuglsang4m(x, cTable)
}
\description{
Effective number of codons based on Fuglsang's formula 4-
}
