% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pminmax.r
\name{pminmax}
\alias{pminmax}
\title{%MinMax of a coding DNA sequence}
\usage{
pminmax(x, z, cut, y = 0, spp = "")
}
\arguments{
\item{x}{a char array of the codons for which %MinMax will be calculated}

\item{z}{the width of the sliding window.}

\item{cut}{a named numeric array of the codon usage table of the host species. Each entry is the number of codons of genes from the reference species, and the name is the case insensitive three letter base char string of the codon.}

\item{y}{a number denoting the codon table to be used. When \emph{cft} is specified, y also needs to be specified. 0 = standard codon table (default), 2 = vertibrate mitchodrial, 3 = ?, 4 = ?. When an option other than those mentioned above is provided, the standard codon table will be used}

\item{spp}{a character string denoting the host species name on which the codon usage table is based. Four reference tables are available: 'ecoli', E. coli; 'insect', insect; 'yeast', yeast; 'celegans', C. elegans. When specified, this option overides \emph{cft}. These build-in codon usage tables are taken from https://www.genscript.com/tools/codon-frequency-table.}
}
\description{
Calculates the %MinMax (Clarke and Clark, 2008) of a DNA sequence coding a protein of interest.
}
\examples{

}
