% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_ndc.r
\name{rm_ndc}
\alias{rm_ndc}
\title{Remove non-degenerate and stop codons}
\usage{
rm_ndc(x, y = 0, keep.stop = TRUE)
}
\arguments{
\item{x}{a list of KZsqns objects}

\item{y}{a vector of numbers denoting the codon table to be used for each KZsqns object. 0 = standard codon table (default), 2 = vertibrate mitchodrial (See dna2aa for additional options). When an option other than those mentioned above is provided, the standard codon table will be used. When the length of this vector is smaller than the number of KZsqns objects supplied, standard codon tables will be used for those remaining KZsqns objects}

\item{keep.stop}{an option indication whether stop codons should also be kept. TRUE: keep stop codons if they are degenerate (default); FALSE: remove stop codons}
}
\description{
Removes non-redundant and stop codons based on the specified codon table. For some compositional analysis, it is advisable to remove non-redundant amino acid coding codons and stop codons.
}
