% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrt.r
\name{rrt}
\alias{rrt}
\title{Random reverse translation}
\usage{
rrt(x, cut, y = 0, spp)
}
\arguments{
\item{x}{a char array of animo acids (e.g., either 'G' or 'Gly' is acceptable but do not mix one letter symbols with three letter symbols in the same array) or codons (e.g., 'GGG'). The first element of the array will be used to determine the type of the input}

\item{cut}{a named numerical array of codon usage table. Case-insensitive three letter codon names are used. Numbers could be the actuall codon counts or scaled ratios}

\item{y}{a number denoting the codon table to be used. When \emph{cut} is specified, \emph{y} also needs to be specified. 0 = standard codon table (default), 2 = vertibrate mitchodrial, 3 = ?, 4 = ?. When an option other than those mentioned above is provided, the standard codon table will be used}

\item{spp}{a char string of the host species name for which the codon usage table will be used. Four build-in reference tables are available: 'ecoli', E. coli; 'insect', insect; 'yeast', yeast; 'celegans', C. elegans. When specified, this option overides \emph{cut}. These build-in codon usage tables are taken from https://www.genscript.com/tools/codon-frequency-table.}
}
\value{
a char array of random reverse translated codons
}
\description{
Randomly reverse translate an animo acid sequence into a DNA sequence according to a codon usage table and a codon table
}
\examples{

}
