% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscu.r
\name{rscu}
\alias{rscu}
\title{Relative synonymous codon usage}
\usage{
rscu(x, y = 0, stop = FALSE)
}
\arguments{
\item{x}{a char array of codons}

\item{y}{an optional number denoting the codon table to be used. 0 = standard codon table (default), 2 = vertibrate mitchodrial (See dna2aa for additional options). When an option other than those mentioned above is provided, the standard codon table will be used}

\item{stop}{a optional boolean denoting whether the RSCU values for the stop codons should be returned. FALSE: NAs will be returned (default); TRUE: the RSCUs will be returned}
}
\value{
a numeric array of RSCUs for each animo acid
}
\description{
Calculates the relative synomynous codon usage (RSCU) for each codon.
}
\details{
The RSCU is calculated for each animo acid coding codon present in the provided codon sequence. When any animo acid is missing from the sequence, each coding codon for that animo acid will have NA returned as their RSCU. The RSCU of stop codons also have NAs in their position as the default setting. To include RSCUs for the stop codons in the returned array, set \emph{stop} as TRUE.
}
\examples{

}
