% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.fasta.r
\name{load.fasta}
\alias{load.fasta}
\title{A function to load fasta files}
\usage{
load.fasta(file = "n", x, ...)
}
\arguments{
\item{file}{Name of the fasta file}

\item{x}{a character array of DNA sequence. Use "A", "T", "G" and "C" to denote the nucleatide composition. When a fasta file is also specified, the fasta file takes priority, and this character sequence is ignored}

\item{...}{additional parameters you might want to pass to the seqinr::read.fasta function}
}
\value{
A list of objects of class KZsqns
}
\description{
Loads the specified fasta file
}
\examples{
dna1 = load.fasta("mitochondrial01.fasta", as.string = TRUE) # not run
}
